drivers: csi: Rework to fix the low framerate issue
Currently, the camera framerate is divided by two. This is because the
driver only puts the full buffer to the queue and calls the callback at
half of the interrupt rate (when there are at least 2 active framebuffers).

This issue is not really visible when using more than two buffers and the
application does not hold the full buffer for too long before submitting
it again to the empty buffers queue. But if we use only two buffers, the
framerate is clearly reduced (by a factor of two).

Especially, when CSI_TransferGetFullBuffer() get called inside the
callback function (as in Zephyr RTOS), framerate is always reduced by a
factor of two despite of the number of buffers used.

Rework the driver to fix these issues as well as to clean up the logics.

Signed-off-by: Phi Bang Nguyen <phibang.nguyen@nxp.com>